requirements.txt
requests
pytz
base58
PyNaCl
colorama
